back/requirements-base.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
requests
django-filter


